+++ packages/kbn-ftr-common-functional-services/services/retry/index.ts
@@ -7,4 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { RetryService } from './retry';
+export { RetryService, type TryWithRetriesOptions } from './retry';
