+++ packages/kbn-ftr-common-functional-services/services/retry/retry_for_success.ts
@@ -92,7 +92,7 @@ export async function retryForSuccess<T>(log: ToolingLog, options: Options<T>) {
     if (lastError && onFailureBlock) {
       const before = await runAttempt(onFailureBlock);
       if ('error' in before) {
-        log.debug(`--- onRetryBlock error: ${before.error.message}`);
+        log.debug(`--- onRetryBlock error: ${before.error.message} - Attempt #: ${attemptCounter}`);
       }
     }
 
@@ -104,9 +104,13 @@ export async function retryForSuccess<T>(log: ToolingLog, options: Options<T>) {
 
     if ('error' in attempt) {
       if (lastError && lastError.message === attempt.error.message) {
-        log.debug(`--- ${methodName} failed again with the same message...`);
+        log.debug(
+          `--- ${methodName} failed again with the same message... - Attempt #: ${attemptCounter}`
+        );
       } else {
-        log.debug(`--- ${methodName} error: ${attempt.error.message}`);
+        log.debug(
+          `--- ${methodName} error: ${attempt.error.message} - Attempt #: ${attemptCounter}`
+        );
       }
 
       lastError = attempt.error;
