+++ x-pack/test_serverless/api_integration/test_suites/common/alerting/rules.ts
@@ -9,28 +9,6 @@ import { RULE_SAVED_OBJECT_TYPE } from '@kbn/alerting-plugin/server';
 import expect from '@kbn/expect';
 import { omit } from 'lodash';
 import { FtrProviderContext } from '../../../ftr_provider_context';
-import {
-  createIndexConnector,
-  createEsQueryRule,
-  disableRule,
-  updateEsQueryRule,
-  runRule,
-  muteRule,
-  enableRule,
-  muteAlert,
-  unmuteRule,
-  createSlackConnector,
-} from './helpers/alerting_api_helper';
-import {
-  createIndex,
-  getDocumentsInIndex,
-  waitForAllTasks,
-  waitForAllTasksIdle,
-  waitForDisabled,
-  waitForDocumentInIndex,
-  waitForExecutionEventLog,
-  waitForNumRuleRuns,
-} from './helpers/alerting_wait_for_helpers';
 import type { InternalRequestHeader, RoleCredentials } from '../../../../shared/services';
 
 export default function ({ getService }: FtrProviderContext) {
@@ -39,7 +17,8 @@ export default function ({ getService }: FtrProviderContext) {
   const esDeleteAllIndices = getService('esDeleteAllIndices');
   const svlCommonApi = getService('svlCommonApi');
   const svlUserManager = getService('svlUserManager');
-  const supertestWithoutAuth = getService('supertestWithoutAuth');
+  const alertingApi = getService('alertingApi');
+
   let roleAdmin: RoleCredentials;
   let internalReqHeader: InternalRequestHeader;
 
@@ -73,19 +52,15 @@ export default function ({ getService }: FtrProviderContext) {
     it('should schedule task, run rule and schedule actions when appropriate', async () => {
       const testStart = new Date();
 
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -130,10 +105,14 @@ export default function ({ getService }: FtrProviderContext) {
       ruleId = createdRule.id;
 
       // Wait for the action to index a document before disabling the alert and waiting for tasks to finish
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
+        retryOptions: {
+          retryCount: 12,
+          retryDelay: 2000,
+        },
       });
       expect(resp.hits.hits.length).to.be(1);
 
@@ -151,7 +130,7 @@ export default function ({ getService }: FtrProviderContext) {
         tags: '',
       });
 
-      const eventLogResp = await waitForExecutionEventLog({
+      const eventLogResp = await alertingApi.helpers.waiting.waitForExecutionEventLog({
         esClient,
         filter: testStart,
         ruleId,
@@ -171,19 +150,15 @@ export default function ({ getService }: FtrProviderContext) {
     it('should pass updated rule params to executor', async () => {
       const testStart = new Date();
 
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -228,10 +203,11 @@ export default function ({ getService }: FtrProviderContext) {
       ruleId = createdRule.id;
 
       // Wait for the action to index a document before disabling the alert and waiting for tasks to finish
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
+        retryOptions: { retryDelay: 800, retryCount: 10 },
       });
       expect(resp.hits.hits.length).to.be(1);
 
@@ -249,13 +225,13 @@ export default function ({ getService }: FtrProviderContext) {
         tags: '',
       });
 
-      await waitForAllTasksIdle({
+      await alertingApi.helpers.waiting.waitForAllTasksIdle({
         esClient,
         filter: testStart,
       });
 
-      await updateEsQueryRule({
-        supertest,
+      await alertingApi.helpers.updateEsQueryRule({
+        roleAuthc: roleAdmin,
         ruleId,
         updates: {
           name: 'def',
@@ -263,15 +239,13 @@ export default function ({ getService }: FtrProviderContext) {
         },
       });
 
-      await runRule({
-        supertestWithoutAuth,
+      await alertingApi.helpers.runRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         ruleId,
       });
 
       // make sure alert info passed to executor is correct
-      const resp2 = await waitForDocumentInIndex({
+      const resp2 = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
@@ -298,18 +272,14 @@ export default function ({ getService }: FtrProviderContext) {
       const testStart = new Date();
 
       // Should fail
-      const createdConnector = await createSlackConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createSlackConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Slack Connector: Alerting API test',
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -341,7 +311,7 @@ export default function ({ getService }: FtrProviderContext) {
       ruleId = createdRule.id;
 
       // Should retry when the the action fails
-      const resp = await waitForAllTasks({
+      const resp = await alertingApi.helpers.waiting.waitForAllTasks({
         esClient,
         filter: testStart,
         taskType: 'actions:.slack',
@@ -353,19 +323,15 @@ export default function ({ getService }: FtrProviderContext) {
     it('should throttle alerts when appropriate', async () => {
       const testStart = new Date();
 
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -407,29 +373,27 @@ export default function ({ getService }: FtrProviderContext) {
       ruleId = createdRule.id;
 
       // Wait until alerts ran at least 3 times before disabling the alert and waiting for tasks to finish
-      await waitForNumRuleRuns({
-        supertestWithoutAuth,
+      await alertingApi.helpers.waitForNumRuleRuns({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         numOfRuns: 3,
         ruleId,
         esClient,
         testStart,
       });
 
-      await disableRule({
-        supertest,
+      await alertingApi.helpers.disableRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
-      await waitForDisabled({
+      await alertingApi.helpers.waiting.waitForDisabled({
         esClient,
         ruleId,
         filter: testStart,
       });
 
       // Ensure actions only executed once
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
@@ -440,19 +404,15 @@ export default function ({ getService }: FtrProviderContext) {
     it('should throttle alerts with throttled action when appropriate', async () => {
       const testStart = new Date();
 
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -498,29 +458,27 @@ export default function ({ getService }: FtrProviderContext) {
       ruleId = createdRule.id;
 
       // Wait until alerts ran at least 3 times before disabling the alert and waiting for tasks to finish
-      await waitForNumRuleRuns({
-        supertestWithoutAuth,
+      await alertingApi.helpers.waitForNumRuleRuns({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         numOfRuns: 3,
         ruleId,
         esClient,
         testStart,
       });
 
-      await disableRule({
-        supertest,
+      await alertingApi.helpers.disableRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
-      await waitForDisabled({
+      await alertingApi.helpers.waiting.waitForDisabled({
         esClient,
         ruleId,
         filter: testStart,
       });
 
       // Ensure actions only executed once
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
@@ -531,19 +489,15 @@ export default function ({ getService }: FtrProviderContext) {
     it('should reset throttle window when not firing and should not throttle when changing groups', async () => {
       const testStart = new Date();
 
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -614,21 +568,21 @@ export default function ({ getService }: FtrProviderContext) {
       ruleId = createdRule.id;
 
       // Wait for the action to index a document
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waiting.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
       });
       expect(resp.hits.hits.length).to.be(1);
 
-      await waitForAllTasksIdle({
+      await alertingApi.helpers.waiting.waitForAllTasksIdle({
         esClient,
         filter: testStart,
       });
 
       // Update the rule to recover
-      await updateEsQueryRule({
-        supertest,
+      await alertingApi.helpers.updateEsQueryRule({
+        roleAuthc: roleAdmin,
         ruleId,
         updates: {
           name: 'never fire',
@@ -645,34 +599,36 @@ export default function ({ getService }: FtrProviderContext) {
         },
       });
 
-      await runRule({
-        supertestWithoutAuth,
+      await alertingApi.helpers.runRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         ruleId,
       });
 
-      const eventLogResp = await waitForExecutionEventLog({
+      const eventLogResp = await alertingApi.helpers.waiting.waitForExecutionEventLog({
         esClient,
         filter: testStart,
         ruleId,
         num: 2,
+        retryOptions: {
+          retryCount: 12,
+          retryDelay: 2000,
+        },
       });
       expect(eventLogResp.hits.hits.length).to.be(2);
 
-      await disableRule({
-        supertest,
+      await alertingApi.helpers.disableRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
-      await waitForDisabled({
+      await alertingApi.helpers.waiting.waitForDisabled({
         esClient,
         ruleId,
         filter: testStart,
       });
 
       // Ensure only 2 actions are executed
-      const resp2 = await waitForDocumentInIndex({
+      const resp2 = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
@@ -683,21 +639,17 @@ export default function ({ getService }: FtrProviderContext) {
 
     it(`shouldn't schedule actions when alert is muted`, async () => {
       const testStart = new Date();
-      await createIndex({ esClient, indexName: ALERT_ACTION_INDEX });
+      await alertingApi.helpers.waiting.createIndex({ esClient, indexName: ALERT_ACTION_INDEX });
 
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         enabled: false,
         consumer: 'alerts',
         name: 'always fire',
@@ -742,41 +694,39 @@ export default function ({ getService }: FtrProviderContext) {
       });
       ruleId = createdRule.id;
 
-      await muteRule({
-        supertest,
+      await alertingApi.helpers.muteRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
-      await enableRule({
-        supertest,
+      await alertingApi.helpers.enableRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
       // Wait until alerts schedule actions twice to ensure actions had a chance to skip
       // execution once before disabling the alert and waiting for tasks to finish
-      await waitForNumRuleRuns({
-        supertestWithoutAuth,
+      await alertingApi.helpers.waitForNumRuleRuns({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         numOfRuns: 2,
         ruleId,
         esClient,
         testStart,
       });
 
-      await disableRule({
-        supertest,
+      await alertingApi.helpers.disableRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
-      await waitForDisabled({
+      await alertingApi.helpers.waiting.waitForDisabled({
         esClient,
         ruleId,
         filter: testStart,
       });
 
       // Should not have executed any action
-      const resp2 = await getDocumentsInIndex({
+      const resp2 = await alertingApi.helpers.waiting.getDocumentsInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
@@ -786,21 +736,17 @@ export default function ({ getService }: FtrProviderContext) {
 
     it(`shouldn't schedule actions when alert instance is muted`, async () => {
       const testStart = new Date();
-      await createIndex({ esClient, indexName: ALERT_ACTION_INDEX });
+      await alertingApi.helpers.waiting.createIndex({ esClient, indexName: ALERT_ACTION_INDEX });
 
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         enabled: false,
         consumer: 'alerts',
         name: 'always fire',
@@ -845,42 +791,40 @@ export default function ({ getService }: FtrProviderContext) {
       });
       ruleId = createdRule.id;
 
-      await muteAlert({
-        supertest,
+      await alertingApi.helpers.muteAlert({
+        roleAuthc: roleAdmin,
         ruleId,
         alertId: 'query matched',
       });
 
-      await enableRule({
-        supertest,
+      await alertingApi.helpers.enableRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
       // Wait until alerts schedule actions twice to ensure actions had a chance to skip
       // execution once before disabling the alert and waiting for tasks to finish
-      await waitForNumRuleRuns({
-        supertestWithoutAuth,
+      await alertingApi.helpers.waitForNumRuleRuns({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         numOfRuns: 2,
         ruleId,
         esClient,
         testStart,
       });
 
-      await disableRule({
-        supertest,
+      await alertingApi.helpers.disableRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
-      await waitForDisabled({
+      await alertingApi.helpers.waiting.waitForDisabled({
         esClient,
         ruleId,
         filter: testStart,
       });
 
       // Should not have executed any action
-      const resp2 = await getDocumentsInIndex({
+      const resp2 = await alertingApi.helpers.waiting.getDocumentsInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
@@ -889,19 +833,15 @@ export default function ({ getService }: FtrProviderContext) {
     });
 
     it(`should unmute all instances when unmuting an alert`, async () => {
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         enabled: false,
         consumer: 'alerts',
         name: 'always fire',
@@ -946,29 +886,29 @@ export default function ({ getService }: FtrProviderContext) {
       });
       ruleId = createdRule.id;
 
-      await muteAlert({
-        supertest,
+      await alertingApi.helpers.muteAlert({
+        roleAuthc: roleAdmin,
         ruleId,
         alertId: 'query matched',
       });
 
-      await muteRule({
-        supertest,
+      await alertingApi.helpers.muteRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
-      await unmuteRule({
-        supertest,
+      await alertingApi.helpers.unmuteRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
-      await enableRule({
-        supertest,
+      await alertingApi.helpers.enableRule({
+        roleAuthc: roleAdmin,
         ruleId,
       });
 
       // Should have one document indexed by the action
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
