+++ x-pack/test_serverless/api_integration/test_suites/common/alerting/summary_actions.ts
@@ -29,25 +29,15 @@ import {
 } from '@kbn/rule-data-utils';
 import { omit, padStart } from 'lodash';
 import { FtrProviderContext } from '../../../ftr_provider_context';
-import { createIndexConnector, createEsQueryRule } from './helpers/alerting_api_helper';
-import {
-  createIndex,
-  getDocumentsInIndex,
-  waitForAlertInIndex,
-  waitForDocumentInIndex,
-} from './helpers/alerting_wait_for_helpers';
-import { InternalRequestHeader, RoleCredentials } from '../../../../shared/services';
+import { RoleCredentials } from '../../../../shared/services';
 
 export default function ({ getService }: FtrProviderContext) {
   const supertest = getService('supertest');
   const esClient = getService('es');
   const esDeleteAllIndices = getService('esDeleteAllIndices');
-
-  const svlCommonApi = getService('svlCommonApi');
   const svlUserManager = getService('svlUserManager');
-  const supertestWithoutAuth = getService('supertestWithoutAuth');
+  const alertingApi = getService('alertingApi');
   let roleAdmin: RoleCredentials;
-  let internalReqHeader: InternalRequestHeader;
 
   describe('Summary actions', function () {
     const RULE_TYPE_ID = '.es-query';
@@ -75,7 +65,6 @@ export default function ({ getService }: FtrProviderContext) {
 
     before(async () => {
       roleAdmin = await svlUserManager.createM2mApiKeyWithRoleScope('admin');
-      internalReqHeader = svlCommonApi.getInternalRequestHeader();
     });
 
     afterEach(async () => {
@@ -98,19 +87,15 @@ export default function ({ getService }: FtrProviderContext) {
 
     it('should schedule actions for summary of alerts per rule run', async () => {
       const testStart = new Date();
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -158,19 +143,27 @@ export default function ({ getService }: FtrProviderContext) {
       });
       ruleId = createdRule.id;
 
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
+        retryOptions: {
+          retryCount: 20,
+          retryDelay: 15_000,
+        },
       });
       expect(resp.hits.hits.length).to.be(1);
 
-      const resp2 = await waitForAlertInIndex({
+      const resp2 = await alertingApi.helpers.waitForAlertInIndex({
         esClient,
         filter: testStart,
         indexName: ALERT_INDEX,
         ruleId,
         num: 1,
+        retryOptions: {
+          retryCount: 20,
+          retryDelay: 15_000,
+        },
       });
       expect(resp2.hits.hits.length).to.be(1);
 
@@ -228,19 +221,15 @@ export default function ({ getService }: FtrProviderContext) {
 
     it('should filter alerts by kql', async () => {
       const testStart = new Date();
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -288,19 +277,27 @@ export default function ({ getService }: FtrProviderContext) {
       });
       ruleId = createdRule.id;
 
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
+        retryOptions: {
+          retryCount: 20,
+          retryDelay: 15_000,
+        },
       });
       expect(resp.hits.hits.length).to.be(1);
 
-      const resp2 = await waitForAlertInIndex({
+      const resp2 = await alertingApi.helpers.waitForAlertInIndex({
         esClient,
         filter: testStart,
         indexName: ALERT_INDEX,
         ruleId,
         num: 1,
+        retryOptions: {
+          retryCount: 20,
+          retryDelay: 15_000,
+        },
       });
       expect(resp2.hits.hits.length).to.be(1);
 
@@ -365,21 +362,17 @@ export default function ({ getService }: FtrProviderContext) {
       const start = `${hour}:${minutes}`;
       const end = `${hour}:${minutes}`;
 
-      await createIndex({ esClient, indexName: ALERT_ACTION_INDEX });
+      await alertingApi.helpers.waiting.createIndex({ esClient, indexName: ALERT_ACTION_INDEX });
 
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -433,7 +426,7 @@ export default function ({ getService }: FtrProviderContext) {
       ruleId = createdRule.id;
 
       // Should not have executed any action
-      const resp = await getDocumentsInIndex({
+      const resp = await alertingApi.helpers.waiting.getDocumentsInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
@@ -443,19 +436,15 @@ export default function ({ getService }: FtrProviderContext) {
 
     it('should schedule actions for summary of alerts on a custom interval', async () => {
       const testStart = new Date();
-      const createdConnector = await createIndexConnector({
-        supertestWithoutAuth,
+      const createdConnector = await alertingApi.helpers.createIndexConnector({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: ALERT_ACTION_INDEX,
       });
       connectorId = createdConnector.id;
 
-      const createdRule = await createEsQueryRule({
-        supertestWithoutAuth,
+      const createdRule = await alertingApi.helpers.createEsQueryRule({
         roleAuthc: roleAdmin,
-        internalReqHeader,
         consumer: 'alerts',
         name: 'always fire',
         ruleTypeId: RULE_TYPE_ID,
@@ -501,20 +490,28 @@ export default function ({ getService }: FtrProviderContext) {
       });
       ruleId = createdRule.id;
 
-      const resp = await waitForDocumentInIndex({
+      const resp = await alertingApi.helpers.waitForDocumentInIndex({
         esClient,
         indexName: ALERT_ACTION_INDEX,
         ruleId,
         num: 2,
         sort: 'asc',
+        retryOptions: {
+          retryCount: 20,
+          retryDelay: 10_000,
+        },
       });
 
-      const resp2 = await waitForAlertInIndex({
+      const resp2 = await alertingApi.helpers.waitForAlertInIndex({
         esClient,
         filter: testStart,
         indexName: ALERT_INDEX,
         ruleId,
         num: 1,
+        retryOptions: {
+          retryCount: 20,
+          retryDelay: 15_000,
+        },
       });
       expect(resp2.hits.hits.length).to.be(1);
