+++ x-pack/test_serverless/functional/test_suites/observability/rules/rules_list.ts
@@ -7,17 +7,7 @@
 
 import { expect } from 'expect';
 import { FtrProviderContext } from '../../../ftr_provider_context';
-import {
-  createAnomalyRule as createRule,
-  disableRule,
-  enableRule,
-  runRule,
-  createIndexConnector,
-  snoozeRule,
-  createLatencyThresholdRule,
-  createEsQueryRule,
-} from '../../../../api_integration/test_suites/common/alerting/helpers/alerting_api_helper';
-import { InternalRequestHeader, RoleCredentials } from '../../../../shared/services';
+import { RoleCredentials } from '../../../../shared/services';
 
 export default ({ getPageObject, getService }: FtrProviderContext) => {
   const svlCommonPage = getPageObject('svlCommonPage');
@@ -31,11 +21,9 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
   const retry = getService('retry');
   const toasts = getService('toasts');
   const log = getService('log');
-  const svlCommonApi = getService('svlCommonApi');
   const svlUserManager = getService('svlUserManager');
-  const supertestWithoutAuth = getService('supertestWithoutAuth');
+  const alertingApi = getService('alertingApi');
   let roleAuthc: RoleCredentials;
-  let internalReqHeader: InternalRequestHeader;
 
   async function refreshRulesList() {
     const existsClearFilter = await testSubjects.exists('rules-list-clear-filter');
@@ -57,10 +45,13 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     numAttempts: number;
   }) {
     for (let i = 0; i < numAttempts; i++) {
-      await runRule({ supertestWithoutAuth, roleAuthc, internalReqHeader, ruleId });
+      await alertingApi.helpers.runRule({ roleAuthc, ruleId });
       await new Promise((resolve) => setTimeout(resolve, intervalMilliseconds));
 
-      await disableRule({ supertest, ruleId });
+      await alertingApi.helpers.disableRule({
+        roleAuthc,
+        ruleId,
+      });
       await new Promise((resolve) => setTimeout(resolve, intervalMilliseconds));
 
       await refreshRulesList();
@@ -68,7 +59,7 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       const rulesStatuses = result.map((item: { status: string }) => item.status);
       if (rulesStatuses.includes('Failed')) return;
 
-      await enableRule({ supertest, ruleId });
+      await alertingApi.helpers.enableRule({ roleAuthc, ruleId });
     }
   }
 
@@ -84,7 +75,6 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
 
     before(async () => {
       roleAuthc = await svlUserManager.createM2mApiKeyWithRoleScope('admin');
-      internalReqHeader = svlCommonApi.getInternalRequestHeader();
       await svlCommonPage.loginWithPrivilegedRole();
       await svlObltNavigation.navigateToLandingPage();
       await svlCommonNavigation.sidenav.clickLink({ text: 'Alerts' });
@@ -107,10 +97,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should create an ES Query Rule and display it when consumer is observability', async () => {
-      const esQuery = await createEsQueryRule({
-        supertestWithoutAuth,
+      const esQuery = await alertingApi.helpers.createEsQueryRule({
         roleAuthc,
-        internalReqHeader,
         name: 'ES Query',
         consumer: 'observability',
         ruleTypeId: '.es-query',
@@ -134,10 +122,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should create an ES Query rule but not display it when consumer is stackAlerts', async () => {
-      const esQuery = await createEsQueryRule({
-        supertestWithoutAuth,
+      const esQuery = await alertingApi.helpers.createEsQueryRule({
         roleAuthc,
-        internalReqHeader,
         name: 'ES Query',
         consumer: 'stackAlerts',
         ruleTypeId: '.es-query',
@@ -159,7 +145,10 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should create and display an APM latency rule', async () => {
-      const apmLatency = await createLatencyThresholdRule({ supertest, name: 'Apm latency' });
+      const apmLatency = await alertingApi.helpers.createLatencyThresholdRule({
+        roleAuthc,
+        name: 'Apm latency',
+      });
       ruleIdList = [apmLatency.id];
 
       await refreshRulesList();
@@ -169,16 +158,16 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should display rules in alphabetical order', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'b',
       });
-      const rule2 = await createRule({
-        supertest,
+      const rule2 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'c',
       });
-      const rule3 = await createRule({
-        supertest,
+      const rule3 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'a',
       });
 
@@ -194,8 +183,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should search for rule', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id];
@@ -215,13 +204,13 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should update rule list on the search clear button click', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'a',
       });
 
-      const rule2 = await createRule({
-        supertest,
+      const rule2 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'b',
         tags: [],
       });
@@ -266,8 +255,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should search for tags', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'a',
         tags: ['tag', 'tagtag', 'taggity tag'],
       });
@@ -289,8 +278,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should display an empty list when search did not return any rules', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id];
@@ -301,8 +290,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should disable single rule', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id];
@@ -329,14 +318,17 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should re-enable single rule', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'a',
       });
 
       ruleIdList = [rule1.id];
 
-      await disableRule({ supertest, ruleId: rule1.id });
+      await alertingApi.helpers.disableRule({
+        roleAuthc,
+        ruleId: rule1.id,
+      });
       await refreshRulesList();
 
       await svlTriggersActionsUI.searchRules(rule1.name);
@@ -360,13 +352,13 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should delete single rule', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'a',
       });
 
-      const rule2 = await createRule({
-        supertest,
+      const rule2 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'b',
       });
 
@@ -392,8 +384,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should disable all selection', async () => {
-      const createdRule1 = await createRule({
-        supertest,
+      const createdRule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [createdRule1.id];
@@ -422,13 +414,16 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should enable all selection', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id];
 
-      await disableRule({ supertest, ruleId: rule1.id });
+      await alertingApi.helpers.disableRule({
+        roleAuthc,
+        ruleId: rule1.id,
+      });
 
       await refreshRulesList();
       await svlTriggersActionsUI.searchRules(rule1.name);
@@ -445,8 +440,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should render percentile column and cells correctly', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id];
@@ -481,8 +476,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should delete all selection', async () => {
-      const createdRule1 = await createRule({
-        supertest,
+      const createdRule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [createdRule1.id];
@@ -508,12 +503,12 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it.skip('should filter rules by the status', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
-      const failedRule = await createRule({
-        supertest,
+      const failedRule = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id, failedRule.id];
@@ -558,8 +553,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it.skip('should display total rules by status and error banner only when exists rules with status error', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       await refreshRulesList();
@@ -582,8 +577,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       );
       expect(alertsErrorBannerWhenNoErrors).toHaveLength(0);
 
-      const failedRule = await createRule({
-        supertest,
+      const failedRule = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id, failedRule.id];
@@ -617,8 +612,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it.skip('Expand error in rules table when there is rule with an error associated', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         name: 'a',
       });
 
@@ -639,8 +634,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       let expandRulesErrorLink = await find.allByCssSelector('[data-test-subj="expandRulesError"]');
       expect(expandRulesErrorLink).toHaveLength(0);
 
-      const failedRule = await createRule({
-        supertest,
+      const failedRule = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id, failedRule.id];
@@ -666,12 +661,12 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should filter rules by the rule type', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
-      const rule2 = await createLatencyThresholdRule({
-        supertest,
+      const rule2 = await alertingApi.helpers.createLatencyThresholdRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id, rule2.id];
@@ -730,36 +725,36 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       };
 
       // Enabled alert
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
-      const disabledRule = await createRule({
-        supertest,
+      const disabledRule = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
-      await disableRule({
-        supertest,
+      await alertingApi.helpers.disableRule({
+        roleAuthc,
         ruleId: disabledRule.id,
       });
 
-      const snoozedRule = await createRule({
-        supertest,
+      const snoozedRule = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
-      await snoozeRule({
-        supertest,
+      await alertingApi.helpers.snoozeRule({
+        roleAuthc,
         ruleId: snoozedRule.id,
       });
 
-      const snoozedAndDisabledRule = await createRule({
-        supertest,
+      const snoozedAndDisabledRule = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
-      await snoozeRule({
-        supertest,
+      await alertingApi.helpers.snoozeRule({
+        roleAuthc,
         ruleId: snoozedAndDisabledRule.id,
       });
-      await disableRule({
-        supertest,
+      await alertingApi.helpers.disableRule({
+        roleAuthc,
         ruleId: snoozedAndDisabledRule.id,
       });
 
@@ -801,28 +796,28 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should filter rules by the tag', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         tags: ['a'],
       });
 
-      const rule2 = await createRule({
-        supertest,
+      const rule2 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         tags: ['b'],
       });
 
-      const rule3 = await createRule({
-        supertest,
+      const rule3 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         tags: ['a', 'b'],
       });
 
-      const rule4 = await createRule({
-        supertest,
+      const rule4 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         tags: ['b', 'c'],
       });
 
-      const rule5 = await createRule({
-        supertest,
+      const rule5 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         tags: ['c'],
       });
 
@@ -864,17 +859,15 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should not prevent rules with action execution capabilities from being edited', async () => {
-      const action = await createIndexConnector({
-        supertestWithoutAuth,
+      const action = await alertingApi.helpers.createIndexConnector({
         roleAuthc,
-        internalReqHeader,
         name: 'Index Connector: Alerting API test',
         indexName: '.alerts-observability.apm.alerts-default',
       });
       expect(action).not.toBe(undefined);
 
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
         actions: [
           {
             group: 'threshold_met',
@@ -902,8 +895,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should allow rules to be snoozed using the right side dropdown', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id];
@@ -922,8 +915,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should allow rules to be snoozed indefinitely using the right side dropdown', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id];
@@ -942,14 +935,14 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     });
 
     it('should allow snoozed rules to be unsnoozed using the right side dropdown', async () => {
-      const rule1 = await createRule({
-        supertest,
+      const rule1 = await alertingApi.helpers.createAnomalyRule({
+        roleAuthc,
       });
 
       ruleIdList = [rule1.id];
 
-      await snoozeRule({
-        supertest,
+      await alertingApi.helpers.snoozeRule({
+        roleAuthc,
         ruleId: rule1.id,
       });
