+++ x-pack/test_serverless/functional/test_suites/search/rules/rule_details.ts
@@ -7,13 +7,8 @@
 
 import { expect } from 'expect';
 import { v4 as uuidv4 } from 'uuid';
-import { InternalRequestHeader, RoleCredentials } from '../../../../shared/services';
+import { RoleCredentials } from '../../../../shared/services';
 import { FtrProviderContext } from '../../../ftr_provider_context';
-import {
-  createEsQueryRule as createRule,
-  createSlackConnector,
-  createIndexConnector,
-} from '../../../../api_integration/test_suites/common/alerting/helpers/alerting_api_helper';
 
 export enum RuleNotifyWhen {
   CHANGE = 'onActionGroupChange',
@@ -34,6 +29,7 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
   const toasts = getService('toasts');
   const comboBox = getService('comboBox');
   const config = getService('config');
+  const alertingApi = getService('alertingApi');
 
   const openFirstRule = async (ruleName: string) => {
     await svlTriggersActionsUI.searchRules(ruleName);
@@ -66,15 +62,11 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
     let ruleIdList: string[];
     let connectorIdList: string[];
 
-    const svlCommonApi = getService('svlCommonApi');
     const svlUserManager = getService('svlUserManager');
-    const supertestWithoutAuth = getService('supertestWithoutAuth');
     let roleAuthc: RoleCredentials;
-    let internalReqHeader: InternalRequestHeader;
 
     before(async () => {
       roleAuthc = await svlUserManager.createM2mApiKeyWithRoleScope('admin');
-      internalReqHeader = svlCommonApi.getInternalRequestHeader();
       await svlCommonPage.loginAsViewer();
     });
 
@@ -88,10 +80,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       const RULE_TYPE_ID = '.es-query';
 
       before(async () => {
-        const rule = await createRule({
-          supertestWithoutAuth,
+        const rule = await alertingApi.helpers.createEsQueryRule({
           roleAuthc,
-          internalReqHeader,
           consumer: 'alerts',
           name: ruleName,
           ruleTypeId: RULE_TYPE_ID,
@@ -261,10 +251,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       const RULE_TYPE_ID = '.es-query';
 
       before(async () => {
-        const rule = await createRule({
-          supertestWithoutAuth,
+        const rule = await alertingApi.helpers.createEsQueryRule({
           roleAuthc,
-          internalReqHeader,
           consumer: 'alerts',
           name: ruleName,
           ruleTypeId: RULE_TYPE_ID,
@@ -369,26 +357,20 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       it('should show and update deleted connectors when there are existing connectors of the same type', async () => {
         const testRunUuid = uuidv4();
 
-        const connector1 = await createSlackConnector({
-          supertestWithoutAuth,
+        const connector1 = await alertingApi.helpers.createSlackConnector({
           roleAuthc,
-          internalReqHeader,
           name: `slack-${testRunUuid}-${0}`,
         });
 
-        const connector2 = await createSlackConnector({
-          supertestWithoutAuth,
+        const connector2 = await alertingApi.helpers.createSlackConnector({
           roleAuthc,
-          internalReqHeader,
           name: `slack-${testRunUuid}-${1}`,
         });
 
         connectorIdList = [connector2.id];
 
-        const rule = await createRule({
-          supertestWithoutAuth,
+        const rule = await alertingApi.helpers.createEsQueryRule({
           roleAuthc,
-          internalReqHeader,
           consumer: 'alerts',
           name: testRunUuid,
           ruleTypeId: RULE_TYPE_ID,
@@ -450,18 +432,14 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
 
       it('should show and update deleted connectors when there are no existing connectors of the same type', async () => {
         const testRunUuid = uuidv4();
-        const connector = await createIndexConnector({
-          supertestWithoutAuth,
+        const connector = await alertingApi.helpers.createIndexConnector({
           roleAuthc,
-          internalReqHeader,
           name: `index-${testRunUuid}-${2}`,
           indexName: ALERT_ACTION_INDEX,
         });
 
-        const rule = await createRule({
-          supertestWithoutAuth,
+        const rule = await alertingApi.helpers.createEsQueryRule({
           roleAuthc,
-          internalReqHeader,
           consumer: 'alerts',
           name: testRunUuid,
           ruleTypeId: RULE_TYPE_ID,
@@ -576,26 +554,20 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
         const testRunUuid = uuidv4();
         const RULE_TYPE_ID = '.es-query';
 
-        const connector1 = await createSlackConnector({
-          supertestWithoutAuth,
+        const connector1 = await alertingApi.helpers.createSlackConnector({
           roleAuthc,
-          internalReqHeader,
           name: `slack-${testRunUuid}-${0}`,
         });
 
-        const connector2 = await createSlackConnector({
-          supertestWithoutAuth,
+        const connector2 = await alertingApi.helpers.createSlackConnector({
           roleAuthc,
-          internalReqHeader,
           name: `slack-${testRunUuid}-${1}`,
         });
 
         connectorIdList = [connector1.id, connector2.id];
 
-        const rule = await createRule({
-          supertestWithoutAuth,
+        const rule = await alertingApi.helpers.createEsQueryRule({
           roleAuthc,
-          internalReqHeader,
           consumer: 'alerts',
           name: `test-rule-${testRunUuid}`,
           ruleTypeId: RULE_TYPE_ID,
@@ -670,10 +642,8 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       });
 
       it('renders a disabled rule details view in app button', async () => {
-        const rule = await createRule({
-          supertestWithoutAuth,
+        const rule = await alertingApi.helpers.createEsQueryRule({
           roleAuthc,
-          internalReqHeader,
           consumer: 'alerts',
           name: ruleName,
           ruleTypeId: RULE_TYPE_ID,
